Speed of texture/tunnel adapts to the ship speed. - removed constant movement per frame + added movement adapted to frame duration
To change the texture/tunnel speed according to the ship speed, the SpeedChangeBehavior of the Ship class needs to know the TextureTransformBehavior of the Tunnel. The changes in Game3D, Ship and Tunnel make this possible. I'm sure this can be done better than introducing all these dependencies, but this quick hack works for the moment.